proj_requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alluka==0.3.1
argcomplete==3.4.0
attrs==23.2.0
boto3==1.34.140
botocore==1.34.140
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorlog==6.8.2
cryptography==42.0.8
Deprecated==1.2.14
distlib==0.3.8
filelock==3.15.4
frozenlist==1.4.1
hikari==2.0.0.dev126
hikari-arc==1.3.4
hikari-miru==4.1.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
nodeenv==1.9.1
nox==2024.4.15
packaging==24.1
platformdirs==4.2.2
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyright==1.1.370
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
ruff==0.5.1
s3transfer==0.10.2
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wrapt==1.16.0
yarl==1.9.4
